refactor LocalIndexesFragment
